+++ Example/NablaExampleApp/NablaExampleApp/FakeAuthenticator.swift
@@ -4,7 +4,7 @@ import NablaCore
 class FakeAuthenticator: SessionTokenProvider {
     static let shared = FakeAuthenticator()
     
-    func provideTokens(forUserId _: UUID, completion: (AuthTokens?) -> Void) {
+    func provideTokens(forUserId _: String, completion: (AuthTokens?) -> Void) {
         // Emulate a call to authenticate the user on your server
         // In your app, you need to replace this with an actual call to your backend to get fresh tokens
         completion(.init(