+++ Sources/NablaCore/Data/DataSources/User/LocalUser.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 struct LocalUser: Codable {
-    let id: UUID
+    let id: String
 }